+++ apps/dialtone-documentation/docs/.vuepress/theme/index.js
@@ -146,7 +146,8 @@ export const dialtoneVuepressTheme = (options) => {
         ]);
       _extractFrontmatter(app, '/guides/content/', options);
       _extractFrontmatter(app, '/components/', options, ['/components/status/']);
-      _extractFrontmatter(app, '/design/', options);
+      _extractFrontmatter(app, '/design/', options, ['/design/colors/usage/', '/design/colors/palette/']);
+      _extractFrontmatter(app, '/design/colors/', options);
       _extractComponentStatus(app);
     },
     // eslint-disable-next-line complexity