+++ packages/dialtone-vue2/components/rich_text_editor/rich_text_editor.vue
@@ -481,6 +481,13 @@ export default {
      * @type {String}
      */
     'selected',
+
+    /**
+     * Event fired when a slash command is selected
+     * @event selected-command
+     * @type {String}
+     */
+    'selected-command',
   ],
 
   data () {
@@ -599,7 +606,12 @@ export default {
       if (this.slashCommandSuggestion) {
         // Add both the suggestion plugin as well as means for user to add suggestion items to the plugin
         const suggestionObject = { ...this.slashCommandSuggestion, ...slashCommandSuggestion };
-        extensions.push(SlashCommandPlugin.configure({ suggestion: suggestionObject }));
+        extensions.push(SlashCommandPlugin.configure({
+          suggestion: suggestionObject,
+          onSelectedCommand: (command) => {
+            this.$emit('selected-command', command);
+          },
+        }));
       }
 
       // Emoji has some interactions with Enter key
@@ -722,6 +734,14 @@ export default {
           attributes: {
             ...this.inputAttrs,
             class: this.inputClass,
+          },
+
+          handleKeyDown: (view, event) => {
+            // When preventTyping is true, only allow backspace to take effect
+            if (this.preventTyping && event.key !== 'Backspace') {
+              return true; // Prevent the event from being processed
+            }
+            return false; // Allow the event to be processed normally
           },
 
           handlePaste: (view, event, slice) => {
@@ -982,12 +1002,6 @@ export default {
       });
       // The content has changed.
       this.editor.on('update', () => {
-        // When preventTyping is true and user wants to type, we revert to last value
-        // If Backspace (keyCode = 8) is pressed, we allow updating the text
-        if (this.preventTyping && this.editor.view?.input?.lastKeyCode !== 8) {
-          this.editor.commands.setContent(this.value, false);
-          return;
-        }
         this.triggerInputChangeEvents();
       });