+++ packages/dialtone-vue3/recipes/conversation_view/feed_item_row/feed_item_row_variants.story.vue
@@ -288,6 +288,7 @@ import { DtAvatar } from '@/components/avatar';
 import { DtIcon } from '@/components/icon';
 import { DtImageViewer } from '@/components/image_viewer';
 import { DtButton } from '@/components/button';
+import { sharedEmojiReactionsData } from '@/recipes/conversation_view/emoji_row/emoji_row.stories.js';
 
 import fryImage from '@/common/assets/fry.gif';
 
@@ -309,28 +310,7 @@ export default {
   data () {
     return {
       fadeState: 'SEARCHED',
-      mockReactions: [
-        {
-          emojiUnicodeOrShortname: '😀',
-          isSelected: true,
-          num: 1,
-        },
-        {
-          emojiUnicodeOrShortname: '😒',
-          isSelected: false,
-          num: 1,
-        },
-        {
-          emojiUnicodeOrShortname: '😌',
-          isSelected: true,
-          num: 99,
-        },
-        {
-          emojiUnicodeOrShortname: '🙃',
-          isSelected: false,
-          num: 99,
-        },
-      ],
+      mockReactions: sharedEmojiReactionsData,
 
       hoverButtons: ['bell', 'living-thing', 'map-pin'],
       persons: ['Jim Halpert', 'Michael Scott', 'Pam'],