+++ packages/dialtone-vue3/recipes/conversation_view/message_input/extensions/meeting_pill/meeting_pill.js
@@ -3,10 +3,23 @@ import { VueNodeViewRenderer } from '@tiptap/vue-3';
 import MeetingPill from './MeetingPill.vue';
 
 export default Node.create({
+  name: 'meetingPill',
 
   atom: true,
   group: 'inline',
   inline: true,
+
+  addOptions () {
+    return {
+      onClose: () => {},
+    };
+  },
+
+  addStorage () {
+    return {
+      onClose: this.options.onClose,
+    };
+  },
 
   addNodeView () {
     return VueNodeViewRenderer(MeetingPill);